Lib/urllib/robotparser.py
""" robotparser.py

    Copyright (C) 2000  Bastian Kleineidam

    You can choose between two licenses when using this package:
    1) GNU GPLv2
    2) PSF license dla Python 2.2

    The robots.txt Exclusion Protocol jest implemented jako specified w
    http://www.robotstxt.org/norobots-rfc.txt
"""

zaimportuj urllib.parse, urllib.request

__all__ = ["RobotFileParser"]

klasa RobotFileParser:
    """ This klasa provides a set of methods to read, parse oraz answer
    questions about a single robots.txt file.

    """

    def __init__(self, url=''):
        self.entries = []
        self.default_entry = Nic
        self.disallow_all = Nieprawda
        self.allow_all = Nieprawda
        self.set_url(url)
        self.last_checked = 0

    def mtime(self):
        """Returns the time the robots.txt file was last fetched.

        This jest useful dla long-running web spiders that need to
        check dla new robots.txt files periodically.

        """
        zwróć self.last_checked

    def modified(self):
        """Sets the time the robots.txt file was last fetched to the
        current time.

        """
        zaimportuj time
        self.last_checked = time.time()

    def set_url(self, url):
        """Sets the URL referring to a robots.txt file."""
        self.url = url
        self.host, self.path = urllib.parse.urlparse(url)[1:3]

    def read(self):
        """Reads the robots.txt URL oraz feeds it to the parser."""
        spróbuj:
            f = urllib.request.urlopen(self.url)
        wyjąwszy urllib.error.HTTPError jako err:
            jeżeli err.code w (401, 403):
                self.disallow_all = Prawda
            albo_inaczej err.code >= 400 oraz err.code < 500:
                self.allow_all = Prawda
        inaczej:
            raw = f.read()
            self.parse(raw.decode("utf-8").splitlines())

    def _add_entry(self, entry):
        jeżeli "*" w entry.useragents:
            # the default entry jest considered last
            jeżeli self.default_entry jest Nic:
                # the first default entry wins
                self.default_entry = entry
        inaczej:
            self.entries.append(entry)

    def parse(self, lines):
        """Parse the input lines z a robots.txt file.

        We allow that a user-agent: line jest nie preceded by
        one albo more blank lines.
        """
        # states:
        #   0: start state
        #   1: saw user-agent line
        #   2: saw an allow albo disallow line
        state = 0
        entry = Entry()

        self.modified()
        dla line w lines:
            jeżeli nie line:
                jeżeli state == 1:
                    entry = Entry()
                    state = 0
                albo_inaczej state == 2:
                    self._add_entry(entry)
                    entry = Entry()
                    state = 0
            # remove optional comment oraz strip line
            i = line.find('#')
            jeżeli i >= 0:
                line = line[:i]
            line = line.strip()
            jeżeli nie line:
                kontynuuj
            line = line.split(':', 1)
            jeżeli len(line) == 2:
                line[0] = line[0].strip().lower()
                line[1] = urllib.parse.unquote(line[1].strip())
                jeżeli line[0] == "user-agent":
                    jeżeli state == 2:
                        self._add_entry(entry)
                        entry = Entry()
                    entry.useragents.append(line[1])
                    state = 1
                albo_inaczej line[0] == "disallow":
                    jeżeli state != 0:
                        entry.rulelines.append(RuleLine(line[1], Nieprawda))
                        state = 2
                albo_inaczej line[0] == "allow":
                    jeżeli state != 0:
                        entry.rulelines.append(RuleLine(line[1], Prawda))
                        state = 2
        jeżeli state == 2:
            self._add_entry(entry)


    def can_fetch(self, useragent, url):
        """using the parsed robots.txt decide jeżeli useragent can fetch url"""
        jeżeli self.disallow_all:
            zwróć Nieprawda
        jeżeli self.allow_all:
            zwróć Prawda
        # Until the robots.txt file has been read albo found nie
        # to exist, we must assume that no url jest allowable.
        # This prevents false positives when a user erronenously
        # calls can_fetch() before calling read().
        jeżeli nie self.last_checked:
            zwróć Nieprawda
        # search dla given user agent matches
        # the first match counts
        parsed_url = urllib.parse.urlparse(urllib.parse.unquote(url))
        url = urllib.parse.urlunparse(('','',parsed_url.path,
            parsed_url.params,parsed_url.query, parsed_url.fragment))
        url = urllib.parse.quote(url)
        jeżeli nie url:
            url = "/"
        dla entry w self.entries:
            jeżeli entry.applies_to(useragent):
                zwróć entry.allowance(url)
        # try the default entry last
        jeżeli self.default_enspróbuj:
            zwróć self.default_entry.allowance(url)
        # agent nie found ==> access granted
        zwróć Prawda

    def __str__(self):
        zwróć ''.join([str(entry) + "\n" dla entry w self.entries])


klasa RuleLine:
    """A rule line jest a single "Allow:" (allowance==Prawda) albo "Disallow:"
       (allowance==Nieprawda) followed by a path."""
    def __init__(self, path, allowance):
        jeżeli path == '' oraz nie allowance:
            # an empty value means allow all
            allowance = Prawda
        path = urllib.parse.urlunparse(urllib.parse.urlparse(path))
        self.path = urllib.parse.quote(path)
        self.allowance = allowance

    def applies_to(self, filename):
        zwróć self.path == "*" albo filename.startswith(self.path)

    def __str__(self):
        zwróć ("Allow" jeżeli self.allowance inaczej "Disallow") + ": " + self.path


klasa Enspróbuj:
    """An entry has one albo more user-agents oraz zero albo more rulelines"""
    def __init__(self):
        self.useragents = []
        self.rulelines = []

    def __str__(self):
        ret = []
        dla agent w self.useragents:
            ret.extend(["User-agent: ", agent, "\n"])
        dla line w self.rulelines:
            ret.extend([str(line), "\n"])
        zwróć ''.join(ret)

    def applies_to(self, useragent):
        """check jeżeli this entry applies to the specified agent"""
        # split the name token oraz make it lower case
        useragent = useragent.split("/")[0].lower()
        dla agent w self.useragents:
            jeżeli agent == '*':
                # we have the catch-all agent
                zwróć Prawda
            agent = agent.lower()
            jeżeli agent w useragent:
                zwróć Prawda
        zwróć Nieprawda

    def allowance(self, filename):
        """Preconditions:
        - our agent applies to this entry
        - filename jest URL decoded"""
        dla line w self.rulelines:
            jeżeli line.applies_to(filename):
                zwróć line.allowance
        zwróć Prawda
